Time: 0 ms (100.00%), Space: 6.1 MB (13.33%) - LeetHub

+++ 0878-nth-magical-number/0878-nth-magical-number.cpp
@@ -3,7 +3,7 @@ public:
     int nthMagicalNumber(int n, int a, int b) {
      long long l = min(a,b);
         long long h = l*n;
-        long long lcm = (a*b)/__gcd(a,b);
+        int lcm = (a*b)/__gcd(a,b);
         long long mid,factor;
         while(l<h){
             mid = l+(h-l)/2;
